// src/models/Enums/EUnits.ts
export enum EUnits{

    UNIT = 'Unit',
    HERO = 'Hero',
    SKELET_ROGUE = 'Skelet-Rogue',
    SKELET_KNIGHT = 'Skelet-Knight',
    SKELET_WIZARD = 'Skelet-Wizard',
    SLIME_KNIGHT = "Slime-Knight",
    GHOUL_ROGUE = "Ghoul-Rogue",
    GHOUL_KNIGHT = "Ghoul-Knight",
    GOBLIN_KNIGHT = "Goblin-Knight",
    GOBLIN_ROGUE = "Goblin-Rogue",
    GOBLIN_WIZARD = "Goblin-Wizard",
    GHOST_BOSS = "Ghost-Boss",
    ELDER_GIANT_BOSS = "Elder-Giant-Boss",
    SKELET_BOSS = "Skelet-Boss"
}